Usages/aliases circular import fix
