+++ bot/__init__.py
@@ -3,5 +3,3 @@ from .converters import *
 from .database import *
 from .send import *
 from .web import *
-from .keep_alive import alive
-from .main import usages, aliases, badarguments
